fix unicrypt - add cache
